+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Ansible Collections for HPE OneView Change Log
 
 ## v6.0.0(unreleased)
-This release extends the planned support of the collections to OneView REST API version 2600 (OneView v6.0)
+This release extends the planned support of the collections to OneView REST API version 2600 (OneView v6.0) and ImageStreamer REST API version 2010 (I3S v6.0) 
 
 #### Bug fixes & Enhancements
 - [#66] (https://github.com/HewlettPackard/oneview-ansible-collection/issues/66) module_utils _merge_connections_boot fails with TypeError
 
 ### Modules supported in this release
+- image_streamer_artifact_bundle
+- image_streamer_artifact_bundle_facts
+- image_streamer_deployment_group_facts
+- image_streamer_deployment_plan
+- image_streamer_deployment_plan_facts
+- image_streamer_os_volume_facts
 - oneview_appliance_configuration_timeconfig_facts
 - oneview_appliance_device_snmp_v1_trap_destination
 - oneview_appliance_device_snmp_v1_trap_destination_facts